Update Driver.cs
This is the diff mod 360 update to the slewtoazimuth code block. It optimises the slew direction such that slews take the shortest distance around the azimuth to reach the target. Git desktop's diff has missed out some of my comments which were there when I subsequently opened the file. Not eassuring.