changes to optimise movement diection for slews
Although I had this sorted in a spreadsheet, the way C#  implements the modulo function is diffeent from  google sheets, so the sheets formula did not work. I had to break it down into three parts - variables for part1, 2 and 3 in the code. A small mod to the c# function method (which I coded from the example on the MS web pages), made it work. Tested on 7-3-19 with steppers motors live.