Bug and defaults
The bug still exists and a workaround is added. The workaround involves try - catch- finally to try to resolve an eror in the Azimuth Get property - pkstepper.transmit("AZ") which does not receive a response from arduino, leaving a hang and the 'response' variable as null.
Other changes set default values for stepper and compass comports, to try to resolve the POTH connection error. Initial testing seems positive, but it is an intermittent error - the best kind.