Implementing setpark using persisten storage
Added a numeric updown control to the setupdialog to store the parkazimuth.
Added internal static double parkazimuth so that it can be accessed in the setupdialog initui(). Writeprofile writes the parkazimuth from the numericupdown to the profile and readprofile successfully reads the parkazimuth value and populates the setupdialog control at runtime. Nice.