Major change
The fact that Dome Azimuth exists as a property which can be accessed has meant that  a fair bit of code was redundant. Particularly in SlewToAzimuth and Slewing and AtPark, which all  prviously Tx'd to get the current Azimuth rather than use the Azimuth Property. Many lines of code are reduced to one e.g. CurrentAzimuth = Azimuth.
In the SlewToAzimuth routine itself, in the routine's name definition, 'Double Azimuth' had to be changed to Double TargetAzimuth to avoid confusion with the Dome.Azimuth property.
In the Azimuth Property, changes are incorporated to provide three retries if Double.Tryparse fails to retun a value for the variable 'az'. The finally clause has been removed.

A lot of code has been deleted and this will need a good TEST before it is used in the Observatory