quite a few bug fixes partcularly relating to shutter open and close
Incorrect use of PKSTEPPER serial object for open and close. Changed to PKCOMPASS.
Shutterstatus code was missing a string terminator '#' causing comms hangups. the string returned from the arduino includes a # mark terminator and this was not removed from the string, causing an error which meant shutterstatus always returned CLOSED as default.