Port change to USB from Radio for Compass
The radio compass is scrapped a) because it was not necessary to have it operate via radio and b) because initiation is troublesome at start of an imaging sess-sion. So code has been changed to include a new USB port called pkShutter. The Compass remains as before pkCompass, but now connection is via USB cable. Needs test.