massive changes in this branch to provide labels which identify which mcu is attached to which com port.
this is really useful in the field as the driver's setup dialog used to present three combo box pick lists, where the user has no idea which one to pick. So now the port is identified to the left of each combo box.
Also the OK button is remodelled so that it only becomes active if all three ports are selected. A lot of work over three days.