moved away from ascom utilities for code which checks for MCU
this is because the transmit function of ASCOM utils was not reliable - it was a pain right from the start.
added a routine which sets the setup( dialog comboports selected item to the ports found. Tested and works.
added a few extra lines into the dome program to help with driver tests.